problem fixed
I found out that the restart button doesn't work for the Level Complete widget. The error is that I set endOverlap of the final gate to the display complete node. The strange thing is that the complete widget is displayed but keyboard input no longer works for the player character but if i change to beginOverlap, it works.